Split everything into Client and Server to support Multiplayer
The idea is that all scene states are stored and processed by the server
(which can be either inprocess/local or out of process remote).  The
client is responsible for rendering and interactions with the scene.
There are two libraries that contain the needed interfaces to impliment
either a client or server.